requirements.txt
firebase-admin
blinker==1.6.2
click==8.1.3
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.3
python-dotenv